pyproject.toml
[tool.poetry]
name = "bank_api"
version = "0.1.0"
description = ""
authors = ["Robert Owen <robert.owen@softwire.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask-restx = "^0.5.0"
pytest-freezegun = "^0.4.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
